+++ nebula/src/clients/meta/MetaClient.h
@@ -637,7 +637,7 @@ class MetaClient : public BaseMetaClient {
       nebula::cpp2::ErrorCode taskErrCode,
       cpp2::StatsItem* statisticItem);
 
-  folly::Future<StatusOr<int64_t>> getWorkerId(std::string ipAddr);
+  folly::Future<StatusOr<int32_t>> getWorkerId(std::string ipAddr);
 
   folly::Future<StatusOr<int64_t>> getSegmentId(int64_t length) override;
